deepfake_detection/requirements.txt
fastapi
uvicorn
torch
torchvision
pillow
requests
opencv-python